代码整理。视频源页使用独立的ViewModel

See merge request android/wls-android!189

+++ module_live/src/main/kotlin/com/walisport/module/live/LiveRemoteManager.kt
@@ -90,7 +90,7 @@ class LiveRemoteManager(private val socketManager: WebSocketManager) {
             apiCode = ApiCode.GET_STANDINGS
         ) {
             Client.MatchTrendReq.newBuilder().apply {
-                this.matchId = matchId.toInt()
+                this.matchId = matchId
             }.build()
         }
         if(result.error == null && result.data != null){
